+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added `tools.BytesToString([]byte)` and `tools.StringToBytes(string)` for fast conversion
+
 ## v3.51.2
 * Added `table/options.ReadFromSnapshot(bool)` option for `session.StreamReadTable()`